perf(files): Speed up query from `get_artifact_bundles_containing_url`
This improves the index and query used in `get_artifact_bundles_containing_url`. The problem at the moment is that we use a `contains`, which can't make use of btree indexes.

Rather than contains, we can use suffix matching. Unfortunately, we still can't make use of an index when writing a query like `url LIKE `%something`. But `url LIKE 'something%'` is fine. So with this in mind, we add an index that reverses and uppers the `url` column, and reverse any values we pass to the query, and switch the query to reverse the url and use `istartswith` for the comparison. This allows us to fully make use of the index.

I also simplified the query a little - we don't really need the joins, and it's cleaner to use exists here. It also works around some problems with filtering to this column.

We tested this in production and in a bundle with 17k indexes the query time went down from 100ms to 5ms. Most folks don't have quite so many rows, but this should improve the costs of these queries in general..